Clear "remote:" even where an output line is shorter than it
In ANSI escape code, the EL code (`CSI n K`) is defined as below:

```
CSI n K : EL – Erase in Line

Erases part of the line. If n is zero (or missing), clear from cursor to the end of the line.
If n is one, clear from cursor to beginning of the line.
If n is two, clear entire line. Cursor position does not change.
```

Here is an examples of `CSI n G` and `CSI n K`.

```
$ echo 'remote: \e[1Gtest'
testte:

$ echo 'remote: \e[1G\e[Ktest'
test
```

Signed-off-by: Ryo Nakamura <r7kamura@gmail.com>